Move WPGraphQL tests + improve CRAP score
