+++ tests/phpunit/Core/LegacyPlayerTest.php
@@ -175,122 +175,6 @@ class LegacyPlayerTest extends WP_UnitTestCase
 
     /**
      * @test
-     * @skip
-     * @dataProvider scriptLoaderTagProvider
-     */
-    public function scriptLoaderTag($postArgs, $tag, $handle, $src, $expect)
-    {
-        global $post;
-
-        set_current_screen('/wp-admin/front');
-
-        $post = self::factory()->post->create_and_get($postArgs);
-
-        setup_postdata($post);
-
-        $output = $this->_instance->scriptLoaderTag($tag, $handle, $src);
-        $output = trim($output);
-
-        if ($expect === true) {
-            $this->assertStringStartsWith('<script id="beyondwords-sdk" type="module">', $output);
-            $this->assertStringContainsString('import BeyondWords from \'https://proxy.beyondwords.io/npm/@beyondwords/beyondwords-audio-player-v2@latest/dist/module/index.js\'', $output);
-            $this->assertStringEndsWith('</script>', $output);
-            $this->assertStringNotContainsString($tag, $output);
-        } else {
-            $this->assertEquals($expect, $output);
-        }
-
-        wp_reset_postdata();
-
-        wp_delete_post($post->ID, true);
-    }
-
-    public function scriptLoaderTagProvider()
-    {
-        $tag    = '<script src="https://example.com/index.js"></script>';
-        $handle = 'beyondwords-sdk';
-        $src    = 'https://proxy.beyondwords.io/npm/@beyondwords/beyondwords-audio-player-v2@latest/dist/module/index.js';
-
-        $invalidTag    = '<p></p>';
-        $invalidHandle = 'some-other-handle';
-
-        return [
-            'invalid handle' => [
-                'postArgs' => [
-                    'post_title' => 'PlayerTest::scriptLoaderTag::1',
-                    'meta_input' => [
-                        'beyondwords_project_id' => BEYONDWORDS_TESTS_PROJECT_ID,
-                        'beyondwords_podcast_id' => BEYONDWORDS_TESTS_CONTENT_ID,
-                    ],
-                ],
-                'tag'      => $tag,
-                'handle'   => 'an-invalid-handle',
-                'src'      => $src,
-                'expect'   => $tag,
-            ],
-            'no post' => [
-                'postArgs' => null,
-                'tag'      => $tag,
-                'handle'   => $handle,
-                'src'      => $src,
-                'expect'   => '',
-            ],
-            'No Content ID' => [
-                'postArgs' => [
-                    'post_title' => 'PlayerTest::scriptLoaderTag::2',
-                    'meta_input' => [
-                        'beyondwords_project_id' => BEYONDWORDS_TESTS_PROJECT_ID,
-                    ],
-                ],
-                'tag'    => $tag,
-                'handle' => $handle,
-                'src'    => $src,
-                'expect' => '',
-            ],
-            'No Project ID' => [
-                'postArgs' => [
-                    'post_title' => 'PlayerTest::scriptLoaderTag::3',
-                    'meta_input' => [
-                        'beyondwords_podcast_id' => BEYONDWORDS_TESTS_CONTENT_ID,
-                    ],
-                ],
-                'tag'    => $tag,
-                'handle' => $handle,
-                'src'    => $src,
-                'expect' => '',
-            ],
-            'Post with everything we need' => [
-                'postArgs' => [
-                    'post_title' => 'PlayerTest::scriptLoaderTag::4',
-                    'meta_input' => [
-                        'beyondwords_project_id' => BEYONDWORDS_TESTS_PROJECT_ID,
-                        'beyondwords_podcast_id' => BEYONDWORDS_TESTS_CONTENT_ID,
-                    ],
-                ],
-                'tag'      => $tag,
-                'handle'   => $handle,
-                'src'      => $src,
-                'expect'   => true,
-            ],
-            'Page with everything we need' => [
-                'postArgs' => [
-                    'post_title' => 'PlayerTest::scriptLoaderTag::5',
-                    'post_type' => 'page',
-                    'meta_input' => [
-                        'beyondwords_project_id' => BEYONDWORDS_TESTS_PROJECT_ID,
-                        'beyondwords_podcast_id' => BEYONDWORDS_TESTS_CONTENT_ID,
-                    ],
-                ],
-                'tag'    => $tag,
-                'handle' => $handle,
-                'src'    => $src,
-                'expect' => true,
-            ],
-        ];
-    }
-
-    /**
-     * @test
      */
     public function jsPlayerParams()
     {
